feat(user-module): :sparkles: reorganize entity and repository paths to use TypeORM structure
