+++ src/modules/user/infrastructure/framework/user.module.ts
@@ -8,9 +8,9 @@ import { DestroyUserController } from '@user/infrastructure/controllers/user/com
 import { UpdateUserController } from '@user/infrastructure/controllers/user/commands/update-user/update-user.controller';
 import { GetUserController } from '@user/infrastructure/controllers/user/queries/get-user/get-user.controller';
 import { ListUserController } from '@user/infrastructure/controllers/user/queries/list-user/list-user.controller';
-import { AccountEntity } from '@user/infrastructure/persistence/entities/account.entity';
-import { ProfileEntity } from '@user/infrastructure/persistence/entities/profile.entity';
-import { UserEntity } from '@user/infrastructure/persistence/entities/user.entity';
+import { AccountEntity } from '@user/infrastructure/persistence/typeorm/entities/account.entity';
+import { ProfileEntity } from '@user/infrastructure/persistence/typeorm/entities/profile.entity';
+import { UserEntity } from '@user/infrastructure/persistence/typeorm/entities/user.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity, AccountEntity, ProfileEntity]), CqrsModule],
